+++ src/pyrepo/commands/begin_dev.py
@@ -4,8 +4,14 @@ from ..util import cpe_no_tb
 
 
 @click.command()
+@click.option(
+    "-N",
+    "--no-next-version",
+    is_flag=True,
+    help="Do not calculate a next version for the project",
+)
 @with_project
 @cpe_no_tb
-def cli(project: Project) -> None:
+def cli(project: Project, no_next_version: bool) -> None:
     """Begin work on the next version of the project"""
-    project.begin_dev()
+    project.begin_dev(not no_next_version)